+++ src/main/java/basemod/animations/AbstractAnimation.java
@@ -3,6 +3,8 @@ package basemod.animations;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g3d.Environment;
 import com.badlogic.gdx.graphics.g3d.ModelBatch;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 
 public abstract class AbstractAnimation
 {
@@ -13,6 +15,12 @@ public abstract class AbstractAnimation
 
     public abstract Type type();
 
-    public void renderSprite(SpriteBatch batch) {}
+    @Deprecated
+    public void renderSprite(SpriteBatch batch)
+    {
+        AbstractPlayer player = AbstractDungeon.player;
+        renderSprite(batch, player.drawX + player.animX, player.drawY + player.animY + AbstractDungeon.sceneOffsetY);
+    }
+    public void renderSprite(SpriteBatch batch, float x, float y) {}
     public void renderModel(ModelBatch batch, Environment env) {}
 }