+++ CHANGELOG.md
@@ -453,3 +453,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 
 #### dev ####
 * Make Spriter animation public (kiooeht)
+* Spriter animations for monsters (kiooeht)